BUG#34134984  MySQL client connection doesn't ask terminal for password
PROBLEM:
In MySQL 8.0.29, login to the client with option -p doesn't prompt for
a password if Option File(eg. ~/.my.cnf) is used.

========================================================================

ANALYSIS:
After a previous bug(BUG#33514253) fixed in 8.0.29, Client does not
prompt for a password if password is already populated. This change was
not required as a part of the fix.

In the case of this bug, since the password is already populated from
Option file, therefore no password prompt is seen when option -p is
used.

========================================================================

FIX:
Client is fixed to prompt for password and use password provided through
command line if option -p is used even if password was populated using
Option File.

Change-Id: I113cf978c17af7cf1f4e71c4b2fe38e0a73ed583