WL#13811 Iterator based EXPLAIN FORMAT=JSON [1/2]
FORMAT=TREE uses ExplainIterator() to generate the EXPLAIN tree output.
With hypergraph optimizer enabled and FORMAT=JSON, follow this same
ExplainIterator() path to generate the new iterator-based JSON output.

Given an AccessPath, convert it into an intermediate Json structure.
This structure can be then used to print it into either json format or
tree format.

ExplainAccessPath() now returns a Json object. This object is passed
to ExplainJsonToString() to convert it into a string.
Explain_format_tree and Explain_format_json have their overridden
ExplainJsonToString() functions.

For supporting the new JSON format, use the same Explain_format_JSON
structure. But have a new enum member m_version which will indicate
whether the JSON format is iterator-based or not.

Materialize and other similar plans need to show their table iterator
plans as their parent plans. So generate a subplan tree for the table
iterator plan and push the main Materialize plan under the leaf node
of the subplan tree.
Note: Earlier, we used to print only the first line of the table
iterator subplan.

Support EXPLAIN ANALYZE. Use iterator->GetProfiler() to collect the
actual costs.

Show top level re-written query in the new JSON format.

SHA Subplan token related changes are done in Explain_format_tree
function. It's a bit different than the original. The subtokens are
generated *while* the tree is traversed for explain output. Earlier,
the whole subplan tree was scanned at each tree node to generate the
token for that node (GetForceSubplanToken was called for each node). A
side effect of this commit is that the list of subplan tokens that is
returned back has a different order (child to parent), but that does
not matter. So an mtr test needed to be changed.

For this commit, the JSON format displays only cost fields. Many more
fields will be included in the next commit.

In --hypergraph mode, many mtr tests had to skip FORMAT=JSON
statements. For any test that has only EXPLAIN statements, the test
itself is skipped.

Change-Id: Ibf5f6c267ebd9c65d1c1e428227c96e88d98ac85