Bug #34353161: Hypergraph: Segfault in ExpandSingleFilterAccessPath
After the fix for Bug#33649667, an extra join condition that was added as a
filter on top of a hash join, could now be moved back to hash join as extra join
conditions (while expanding the filter paths). In doing so, it also moves
conditions having subqueries to the hash join. This leads to problems when
the subquery needs to be materialized. For a subquery that needs to be
materialized, currently a filter access path is always inserted. However
for the failing case, moving the condition with subquery to the hash join
leaves no condition behind to create a filter access path. This leads to the
failing asserts where a filter access path is always expected.

Solution is to not add conditions with subqueries to the hash join
conditions. But instead leave them as filters on top of the hash join
itself.

Change-Id: I7a3148a0563a1b5950e4fefbc2cb231693077d67