Bug#33876690 - Server may exit when creating views with a chain of dependencies
Issue:
The stack might get exhausted while creating a nested view which
in turn references a nested view.

This is because the SELECT nesting level limit of 64 is not
enforced while preparing a view.

Fix:
Check the nesting level while preparing a view and throw an
error if limit exceedes.

Change-Id: I4eef53af614d6c2be337e2cba32cdd27c39487ad