Bug#34347644: Hypergraph Performance Issue : TPCDS q64 Hypergraph               optimizer is ignoring the STRAIGHT_JOIN HINT
Contrary to what the bug title says, hypergraph is not ignoring the
straight_join hint. Since straight joins are associative,
it still produces a valid plan as per the directives.
However, it does not pick the ideal plan. When hints are
not present, it does indeed pick a much better plan.
But, for heatwave this is a problem because of the number
of partial plans that are generated.
The problem when straight join hints are given is that only a
few combinations are possible. And with these limited
combinations and the order in which the join conditions
get pushed down to each of these joins, makes the combination
for the best join not possible for this particular case.
E.g.
For a query of this form (SJ - straight_join):
t1 SJ t2 SJ t3 SJ t4 with 4 join conditions:
t3->t4 has a non equality join condition
t2->t4 has an equijoin condition
t1->t3 has an equijoin condition
t1->t2 has an equijoin condition
Call to optimize_cond() places all the equalities at the end of
the final where/join condition. So the in-equality is placed first
in this case.
When optimizer is trying to push the conditions down to joins
initially, it sees the non equality join condition first and
pushes it down for a join between t3->t4 (2 large tables in case
of the non-performant TPC-DS query).
Then it sees the (t2->t4) equi-join condition. It concludes that
it's better to place this condition for the join between (t2)->(t3t4)
because of 2 reasons. The first one being that the used tables
for the join t3->t4 has t3,t4 because of the already present join
condition and the second one being that t3 needs to be seen
before t4 (because of stright_join).
For the next join condition between (t1->t3), it places it on the
join between (t1,(t2t3t4)) and the same with the last condition.
The same reasoning persists when creating edges resulting
in the join between (t3,t4) which are large tables in the actual
use case.
However, if the optimizer were to see the equalities first,
it would place the join condition in such a way that the following
join was possible ((t1 HJ t2) HJ t3) HJ t4) which is still valid as
per the hints but a much better join.
The same problem does not happen when the hints are not used
because many other joins are possible and therefore it could create
more edges.
It could be argued that optimizer should have been able to figure
out the valid edge between t1->t3. However its easier to fix this
particular case by just sorting conditions in a way that the
equalities are seen earlier. So we take the easier route for now.
And putting equijoin conditions ahead of others does no harm.
Also that at this point, the predicates are not ordered in any way
except for optimize_cond() doing a bit of sorting for the multiple
equalties. So in this patch, we put all the equijoin conditions
at the begining followed by other types of conditions. We also make
sure that the sorting that was done while creating hash join
conditions is taken care here itself i.e putting all the conditions
with subqueries and the expensive conditions at the end of the list.
This results in a few failing unit tests. The necessary changes have
been done for those tests.

Change-Id: I6880ffbe6b3fa177a3fbb7612aed3f46eb6440d7