Bug #34370933 Assertion `max_length == 4' failed in MySQL 8.0.29
We hit an assertion when creating fields of type YEAR for temporary
tables holding UNION results.

In set_data_type_year(): set length to four, and set the unsigned bit.

Field_year is an unsigned tinyint of length four, so let the functions
year() and cast_as_year() have the same type.

Change-Id: I485b8ac8559b0dbf25d25e94a24223006adaecfd