Implement logic to be able to get a consistent GTID
This adds logic so that with a `START TRANSACTION WITH CONSISTENT
SNAPSHOT`, we can also get a correct GTID that is appropriate at the
time the transaction is started.

It locks around the logic when a consistent snapshot is created in the
underlying storage engine and grabs the current GTID under lock at that
point.

That means that no other threads can update the GTID at the exact same
point in time avoiding the race that currently exists if you execute
`SELECT @@global.gtid_executed` which doesn't enforce such locking.

The locking overhead is should be relatively small. When using a
consistent read snapshots, systems (like Vitess) often want to get a
consistent GTID. Today that's only possible by then also locking all
tables which is much more expensive.

Signed-off-by: Dirkjan Bussink <d.bussink@gmail.com>