Bug #34303841: Inappropriate error handling for max_allowed_packet
The MYSQL error was not reset to ER_SERVER_LOST when a prior error
caused a disconnect (e.g. ER_NET_PACKET_TOO_LARGE) and C API
reconnect is disabled.
mysql.cc doesn't use libmysql's reconnect option but instead does
its own. But to trigger this it needs to get an error that either
CR_SERVER_LOST or CR_SERVER_GONE_ERROR. And it was getting
ER_NET_PACKET_TOO_LARGE repeatedly instead despite it calling
mysql_real_query that is supposed to reset the error.

Fixed by reverting the removal of the reset done by wl 12999.

Also homogenized the error that it's reset to to always be
CR_SERVER_LOST. And updated an existing bugfix test to test for the
new error.

Test case added.

Change-Id: Ideca371a258349b7438a8d65412d4f16fb5496c2