Bug#34402003: HYPERGRAPH BUG: Offload issue with execute statement.
Hypergraph was using the non optimized join condition in case of prepared
statements because of which multiple equalities were not looked into
resulting in not offloading the query to secondary engine.
During JOIN::optimize, a copy of the original join condition is used for
optimization in optimize_cond() for prepared statements. This condition
needs to be used instead of the original condition for prepared statements.
This was not a problem for non-prepared statements because there is no
copy made in that case. The same join condition is used for optimization
as well.
There is also an additional change while creating equalities from multiple
equalities(noticed after this change). For a multiple equality that is
established to be always false, "Item_equal" sets "cond_false" to true.
It is expected that this is checked before creating simple equalities.
This is by design because a multiple equality cannot have two constant
items.
For ex: if we have f1 = f2, f1 =5 , f2 = 10, here it first creates
multi_equal(f1,f2), then adds (5,f1,f2) and when it looks at "10", it concludes
this will always evaluate to "false" and therefore sets "cond_false" to true.
However if this flag is not checked, then the resulting simple equalties
would look like this - f1=5, f2=5. This is not correct. So we need the
additional check for "false" conditions.

Change-Id: I9103fa41a185f23db5bec0d1fc24c973a602e333