Bug #34380739 Make ndbxfrm tool be able to show header and trailer for NDBXFRM1 file
Added option --detailed-info to ndbxfrm to print file header and trailer.
It is complementary to old version --info that now can be used just to
print file compression and encryption flags.

Added a private method 'open' to the ndbxfrm_file class used internally by:
- (public) 'open' method that should be used if we need to decrypt the file
contents. It requires a valid key/password.
- (public) print_header_and_trailer method used by ndbxfrm tool to read and
print the file header and trailer without decrypting the file payload.
In this case a valid key/pwd is not necessary.

Change-Id: Ic00cc387ab8df02cd4129160d6402beae4bbb082