Bug#34417282 No program to inspect secrets file, extract key for manual usage of tools.
Tool to inspect secrets file and extract node master key.
The tool receives as argument the filesystem password and an
existing secrets file. It uses ndbxfrm_file to open, decrypt
and read the file contents.
First 8 read bytes are used to verify if it is a valid secrets file.
Node Master Key length (4 bytes) and Node Master Key (32 bytes) are
then extract and KEY is printed to stdout in hex format.

Change-Id: I89464ea203f3cc9704a1ec7a954ab6b9d7be3203