Bug #33996054 Result mismatch with input column re-order for GREATEST()
The problem was in type-resolving of expressions:
  greatest(f1, f2, f3, ...)
where some of the arguments are time-related (date, time, timestamp ..)
and other arguments are not.

The output was dependent on the permutation of (f1, f2, f3, ...).

This patch introduces a "ranking" of the various time-related types,
so that we get a consistent result, regardless of argument permutation.

Change-Id: I4a5b087e203c724e092c230c7df2887b0a5ea200
(cherry picked from commit 1f63183e93c93d7858b18fdcf2befb1fb0a57f65)