BUG#33818238: Signal 11 occured while ran an insert statement.
Through code analysis it was observed that
`MYSQL_BIN_LOG::ordered_commit()` was calling
`update_binlog_end_pos(binlog_file, pos);` when the `binlog_file` and
`pos` could be invalid. This situation is unlikely since that
should be considered a flush or sync error, but the implementation
was not considering that hypothesis which could cause unpredictable
results.

To solve the above issue, the implementation now only calls
`update_binlog_end_pos(binlog_file, pos);` when `binlog_file` and
`pos` parameters are valid.

Change-Id: I7d7921eb80181cf632170746fa4a27ad86fef985