Bug#34448749: Hypergraph Crash : FindBestQueryPlan
ApplyDistinctAndOrder() attempts to access the first candidate plan
without checking that there in fact is a candidate plan. This fails if
the list of candidate plans is empty. It can be empty if a secondary
engine has rejected all the candidates in the previous step.

Fixed by checking if all plans have been rejected and returning early
in that case, without attempting to inspect the first plan in the
list.

Change-Id: I6b9164b6d586dc3c72b9e9b35a06e0d634fe563b