Bug #32819101 	ASSERTION FAILURE: FIL0FIL.CC:12089:FIL_SYSTEM->SHARD_BY_ID(ID)->MUTEX_OWNED()
During recovery in case when a table was opened for recovery during recv hash apply and then in a next redo block scan a space deletion is found, the space deletion is redone and `fil_space_free ` is called without fil_shard latch. It is OK, but the debug assertion fails in `has_no_references`. Assertion is updated to check for mutex or the space being private - the reference count can't be bumped if there is no way to get a reference to it from fil_shard.

Change-Id: Ie472b230fe2b5b035721ae3d6f51572ab37a5b36