Bug#34333017: Hypergraph Performance Issue : TPCH q21 SF 256
For the query in bugpage, one of the join order is not proposed
because a conflict rule prevents it from joining the said tables.
This conflict rule is placed because currently it is using
table_map of the tables used in the join condition to create the
rule when it is expected to be using the NodeMap for these tables.
So we correct that in this patch to use NodeMap.

Change-Id: I13f31ca62dd25ab6d4c1062a734b6e174b2c4ee5