WL14955: Have backtraces of failed runs appear in performance_schema.error_>
To make sure no potentially valuable information is lost, the
MySQL server redirects the default streams (stdout/stderr) to
the "traditional" error log-file.

This WL aims to check at start-up whether the last item in the
traditional error-log is a stackdump, and if so, to make that
stackdump available in the performance_schema table, error_log.
This functionality is available even when the traditional log
is not enabled as a log-sink in @@global.log_error_services.

Change-Id: Ia488ee19715bb58254f78b97c0a189a282591fa2