pb2 fix for main.slow_log with --ps-protocol (for testing added by Bug#33732907).
This calls back to the fact that when running in "normal mode",
we see one COM_QUERY row for each statement, whereas when running
with --ps-protocol, we see a COM_STMT_PREPARE/COM_STMT_EXECUTE
pair for each statement. This affects counters and log file/table
contents.

This test now accounts for the mode it runs in and adjusts row
counts accordingly. Since we no longer show individual log lines
in the result, we now take separate counts of lines of interest
with and lines of interest without obfuscated passwords.

Change-Id: Id267a87f4ae1df835a2974dfc405d9aeee0ca8fe