Bug #33781442 BIT column serialized incorrectly for aggregate function
An aggregate query on a BIT field returns a value which is formatted
as a bit string, but has the BINARY flag. This is because the
BINARY flag is automatically added to all result items which have a
binary charset. This should not be done for BIT results since they
are always formatted as bit strings.

This is fixed by adding a check to skip setting the BINARY flag for
BIT results.

Change-Id: If73f511cb27a4e63436bf7e48305bf7b7d197859