Bug#34448531: Hypergraph Offload Issue : Join condition marked as extra predicate
A condition like t1.a = t2.a + t3.a is not considered an equi-join condition
because the right side of the condition has more than one table. As a result,
this makes it to the extra predicate of the hash join. However, if the
optimizer knows that "t2.a = t3.a", the above condition could well be
transformed into t1.a = t2.a + t2.a which could be used as an equi-join
condition. This is done in the old optimizer after the join order is decided.
Hence heatwave is able to offload in case of the old optimizer and not in
Hypergraph.
Solution is to propagate such equalities into all conditions that are not
multiple equalities. This results in several unit tests failing because some
of these fields in non-equalities were getting replaced with another table's
field or the conditions were written in such a way that optimizer could now
detect that these conditions are going be always false. E.g. t1.x = t2.x
and t1.x <> t2.x would now be transformed into t1.x = t2.x and t1.x <> t1.x.
To retain the original intention of the test case, all such cases are changed
to using different field from the same table.

Change-Id: Iaf9c8b393bb563867c6ac56bc637804780648269