Bug#34482783 NDB: mysqld crash in ::is_pushable_within_nest()
When the NDB join pushdown handler finds a table which can't be
pushed down, it tries to produce an explain-string for the reject
reason - This often includes fetching the table names as well.

In some cases the optimizer might have entirely optimized away
the table access. For this particular bug the optimizer finds
that the join condition (IS NULL on a non-nullable column) is
such that no rows will ever be returned. Thus a 'zero-rows' AccessPath
will replace the original table acecss.

This results in that the TABLE* can't be looked up any more from
the NDB handler and a nullptr was accessed when formating the sttring.

This patch introduce a check for 'nullptr' and prints a more generic
explain-string, not including the table name if no table was found.

Note that there are ~hundreds of other explain printing places
in this code as well, where the possible nullptr returning get_table(()
is called. However, except for the places fixed in this patch, we
call get_table() with the current 'table->', 'parent->' or
'm_join_root->', which is already known to be non-nullptr, in these places.

Change-Id: Icd43449f90cc70b8557cb5be9d8f5d9b5bacbe33