WL#15407: NDB Replica updates filter
The NDB cluster plugin can now configure a filter chain in the NDB
data nodes that prevent unnecessary receival of irrelevant data
events, saving network I/O and its associated work. Two filters are
implemented in this work: one for changes that would never be written
to the binary log (always ON), and another that is dependent on the
log-replica-updates option.

Problem:

Upon data updates in the NDB nodes, these send an event describing the
change to the subscribing API nodes. These API nodes, generally the
MySQL server, use the event to extract what happened in the data nodes
and operate on it. The NDB plugin on MySQL server acts on this
received data using the binlog injector thread, by writing to the
binlog. However, there are some MySQL servers that do not want to
write certain changes on the binlog; e.g., changes that were applied
by a replica MySQL server. For this case, MySQL server has the
`log-replica-updates` option that controls if a change applied by a
replica should be logged to the server's binlog.

- If `log-replica-updates` is ON, then NDB cluster plugin discards
replica applied changes. But, this wastes bandwidth and the associated
resources.

Analysis:

The NDB data nodes use the SUMA block to handle the subscribers and
send the event containing the updates that occurred. The subscription
is created and started on the setup of the event stream for any given
table that is binlogged. This is the responsibility of the binlog
thread on the NDB plugin, when setting up a table to be
monitored.

- If the binlog thread can push the `log-replica-updates` filtering to
the data nodes, then the event sending is skipped and thus bandwidth
is reduced.

Solution:

On setting up the signal to start the event
subscription (`SubStartReq`), the NDB API can include some additional
options (extending the signal words). On receipt of the signal, the
SUMA block configures the subscriber by setting up a filter to apply
on the table event data sending loop.

This option setup as a filter, which is essentially a 32bit value
containing the multiple modes of filtering updates that the subscriber
wants; currently only concerned with two, log-replica-updates=OFF and
log_bin=OFF.

The filter for `log-replica-updates=OFF` operates by checking the
`anyValue` of the data to broadcast to subscribers, after a given
I/U/D update. The `anyValue` (documented at
storage/ndb/plugin/ndb_anyvalue.cc) can contain the server-id that
applied the change. If it is different than 0 and not a reserved
value (refer to the `anyValue` documentation) then it contains the
server-id. If the condition of anyValue == server-id &&
log-replica-updates=OFF applies, then the event is not sent.

The filter for `log-bin=OFF` operates by checking if the `anyValue`
contains the reserved value of no-logging (0x8000007f). Any change
that is not to be binlogged on any MySQL server contains the
no-logging code, therefore it is uneecessary to send to subscribing
nodes as it would be rejected anyhow.

Change-Id: Ifc8596ff77764bfa7549a56336365eb2aaa516ae