Bug #34302461 EXPLAIN cost and row estimates are wrong for UNION [1/2] Bug #33834146 'EXPLAIN ANALYZE' cost + elapsed time not cumulative when using materialization
This patch is a precursor for the actual fix of the above bugs. The
patch makes AccessPath.num_output_rows, and adds a getter and a setter
for it. (This was useful when tracking down the cause of those bugs,
and will hopefully aid in analyzing other bugs in the same area.)

Change-Id: Iff8543eb55df76f4351cd31f532df8f11f86f72f