Bug#34311472: Assertion in val_real() failed in MySQL 8.0.29
An assertion failed in debug builds in certain queries performing
full-text search in boolean mode on non-string columns in MyISAM
tables. No problems were seen in release builds.

Item_func_match::val_real() has an assertion to verify that all the
items in the MATCH clause are references to base columns. This is
ensured by checks during resolving. However, constant propagation may
replace the column references by constants after resolving, and then
the assertions fail.

This can only happen with MyISAM. InnoDB requires that the columns
must have FULLTEXT indexes, and FULLTEXT indexes can only be added to
string columns, and constant propagation does not propagate string
constants. MyISAM, on the other hand, allows MATCH columns without
indexes in boolean mode, and therefore also allows integral columns,
which are subject to constant propagation.

Full-text search on non-strings is an uncommon corner case which is
probably not worth optimizing for. We therefore disable constant
propagation on the MATCH clause, so that we can safely assume that it
only contains column references.

Change-Id: Iae7ba3aad3c697075119fcc9b9b9b00dc8ba31f8