BUG#34464526: Missing mysql_start_failover_channels_if_primary after upgrade to 8.0.27+
From MySQL 8.0.26, Group Replication has the capability to set
actions for the members of a group to take in specified situations.
The only existing member action was
`mysql_disable_super_read_only_if_primary`.
On 8.0.27 the member action
`mysql_start_failover_channels_if_primary` was added.

When a member joins a group, it replaces its member actions
configuration by the one of the group.
There is one case on which this can create issues, when a 8.0.27+
server joins a 8.0.26 group.
On that case the group will have the member actions:
 * mysql_disable_super_read_only_if_primary

The 8.0.27+ server before joining the group will have:
 * mysql_disable_super_read_only_if_primary
 * mysql_start_failover_channels_if_primary
though after joining the group, it will have:
 * mysql_disable_super_read_only_if_primary
After all group members are upgraded on 8.0.27+ and rejoined the
group, the action `mysql_start_failover_channels_if_primary` is
lost.

To solve the above issue, now a server joining a group, after
replacing its member actions configuration will check if the member
action `mysql_start_failover_channels_if_primary` does exist, if not
its default configuration is added.

Change-Id: I4d6183eaf14917ce7031deafb160527dcda575ab