Bug#34486874 Error 4829 'FirstInner/Upper has to be an ancestor or a sibling' from NDBCLUSTER
Preparation patch: Provide a stable set of m_uppper_nest vs
m_ancestors_nest tables.

Setting up the set of possible parents/ancestors of a pushed table partly
on which upper_nests each particular table is 'inside'
(i.e. 'is outer-joined with') as well as which tables are being refered
from the join conditions.

When the join-pushdown handler prepares the pushed joins, it keep track
of the set of upper tables in the bitmap 'm_upper_nest', as well as its
own m_inner_nest tables (i.e. The tables it is inner-joined with).
These sets of tables are collected from the query-plan provided by
the optimizer, as represented in the AccessPath structures.

In most cases the join conditions will only refer tables from its
upper_nests and/or inner_nests. E.g. in the join-plans:

( '(...)' represents an inner_nest of tables, where its embedding nests
are the upper_nests)

    t1, (t2), (t3, t4)

  -> t2 outer joins with t1
  -> t3 outer joins with t1
  -> t4 inner joins with t3

We will usually find:

  - t4's join condition refers t3 in its inner_nest, or the upper table t1
  - t3's join condition refers the upper table t1.
  - t2's join condition refers the upper table t1.

However, the optimizer may provide 'plans' where either t3 or t4 refers t2 column.
(Often a result of optimize rewrites of the query) That is, we are refering tables
not in our m_upper_nests !!

We used to detect and handle such out-of-embedded_nests-references in
the method ndb_pushed_builder_ctx::is_outer_nests_referable().
This method checked if such references conflicted with SPJ limitation,
and if accepted the m_upper_nests were *modified* to include the
ancestors actually refered. Thus, in such cases we could end up with
[t1,t2] being the upper_nests of t3 and t4.

As this modification happened midway during pushed join handling, we
might have some logic based on that [t1] was the (original) upper_nest
of t3 and t4, and some based on the upper_nests [t1,t2].

This patch introduce the additional bitmap m_ancestor_nests, as well
as the method ::ancestor_nests(). The m_upper_nests will be kept
'static', such that they always represents the actuall join-plan
topoligy as provided by the AccessPath.

The ancestor_nests will represents the collection of nests which
the join conditions actually refers into. In 'normal' cases the
m_ancestor_nests will be the same as the m_upper_nest. However in
such out-of-embedded_nests-references as discussed above, the
m_ancestor_nests will be set up to contain [t1,t2], while the
m_upper_nests will be kept 'stable' at [t1].

The method is_outer_nests_referable(), which used to change
the m_outer_nests, is renamed to set_ancestor_nests(). This
is were the ancestor_nests now are set up. The preceeding
logic for calling into this method has been modified as we now always
need to call this method - not only in the unusual cases where outer_nests
had to be modified. There are remaining code cleanup in this area left for
a later patch.

Note that most previous usage of m_outer_nests were through
the embedding_nests() method, which returned 'outer_nest+inner_nest'.
This method has now been accompanied by ancestor_nests() returning
ancestor_nests+inner_nest.

Previous usage of embedding_nests() has been cleaned up such that
we still call embedding_nests() where we need to check the actuall
'embedding' as defined by the AccessPath, and change to using ancestor_nests
when checking conflicting table/nest references in the join condition
dependencies.

No query plan changes are expected as result of this patch !!!

Change-Id: Ie4b9011e1c90dba7942f6cefc8520ec304ce3fd7