Bug#34494609: Hypergraph Offload issue : Filter condition not compatible with HeatWave
The hypergraph optimizer was able to detect an always false condition
in the WHERE clause during normalization of the conditions, but
instead of pruning away the entire query, it returned a query plan for
a join with a post-join filter that filtered out all the rows returned
by the join. This is an inefficient plan.

This patch makes the hypergraph optimizer prune away the entire query
in such cases, and only return a ZERO_ROWS access path (or
ZERO_ROWS_AGGREGATED in case of implicitly grouped queries) instead of
a join plan.

Change-Id: Id8bf6d6d014cf4eb85e12f08c01db6681543fc7e