Bug#34418412: regression in 8.0.29: crash in copy_field_info
A constant expression used in the select list is referenced
in the having condition inside a derived table. This reference
is used in a where clause outside the derived table which is
valid for derived condition pushdown. While pushing the condition
down, a clone of the expression is made. While the original
expression has an Item_ref to a constant, the cloned expression
is created as a field item. This results in the mismatch while
copying field information leading to server exit.
Solution is to make Item_ref::print() print the reference if
its a constant item.

Change-Id: Iec99ba11f0e9d153fa54603377ea81d53635b4dd