Bug #34302461 EXPLAIN cost and row estimates are wrong for UNION [2/2] Bug #33834146 'EXPLAIN ANALYZE' cost + elapsed time not cumulative when using materialization
Codewise, the path for accessing a materialized query is a child of the
MATERIALIZE AccessPath object (as member table_path). But in EXPLAIN
output, we reverse this order, e.g.:

-> Table scan on d1  (cost=2.51..2.51 rows=1)
   -> Materialize  (cost=2.96..2.96 rows=1)

This is confusing, as cost is supposed to be cumulative, so that the
cost of a given access path should comprise the cost of executing the
path plus the cost of any descendant paths.

This patch corrects that, so that the displayed cost for 'Materialize'
will be the cost of the subquery that is materialized plus the cost of
materialization.  And the cost of table_path ('Table scan' in the
example above) will be the cost of 'Materialize' plus the cost of
scanning the table.

In some cases, table_path does not point directly to the AccessPath accessing
the materialized table. Instead, there is a number of intermediate AccessPath
objects, e.g.:

MATERIALIZE->SORT->FILTER->TABLE_SCAN

The function MoveCompositeIteratorsFromTablePath() reshuffles this into

SORT->FILTER->MATERIALIZE->TABLE_SCAN

This patch ensures that the cost of those intermediate paths (SORT and FILTER
in the example above) are also recalculated so that they are cumulative.

The patch also ensures that we get a row estimate for union materialization
into anonymous tables. (This was always zero.)

Change-Id: I1dd1117d5f1c38b10c66bba559737c2a2d1b04ab